+++ jupiter-transport/jupiter-transport-netty/src/main/java/org/jupiter/transport/netty/NettyTcpConnector.java
@@ -21,10 +21,12 @@ import io.netty.channel.ChannelOption;
 import io.netty.channel.EventLoopGroup;
 import io.netty.channel.WriteBufferWaterMark;
 import io.netty.channel.epoll.EpollEventLoopGroup;
+import io.netty.channel.kqueue.KQueueEventLoopGroup;
 import io.netty.channel.nio.NioEventLoopGroup;
-import org.jupiter.transport.UnresolvedAddress;
-import org.jupiter.transport.JConnection;
+import org.jupiter.common.util.JConstants;
 import org.jupiter.transport.JConfig;
+import org.jupiter.transport.JConnection;
+import org.jupiter.transport.UnresolvedAddress;
 
 import java.util.concurrent.ThreadFactory;
 
@@ -36,30 +38,30 @@ import java.util.concurrent.ThreadFactory;
  */
 public abstract class NettyTcpConnector extends NettyConnector {
 
-    private final boolean nativeEt; // Use native epoll ET
+    private final boolean isNative; // use native transport
     private final NettyConfig.NettyTcpConfigGroup.ChildConfig childConfig = new NettyConfig.NettyTcpConfigGroup.ChildConfig();
 
     public NettyTcpConnector() {
         super(Protocol.TCP);
-        nativeEt = true;
+        isNative = true;
         init();
     }
 
-    public NettyTcpConnector(boolean nativeEt) {
+    public NettyTcpConnector(boolean isNative) {
         super(Protocol.TCP);
-        this.nativeEt = nativeEt;
+        this.isNative = isNative;
         init();
     }
 
     public NettyTcpConnector(int nWorkers) {
         super(Protocol.TCP, nWorkers);
-        nativeEt = true;
+        isNative = true;
         init();
     }
 
-    public NettyTcpConnector(int nWorkers, boolean nativeEt) {
+    public NettyTcpConnector(int nWorkers, boolean isNative) {
         super(Protocol.TCP, nWorkers);
-        this.nativeEt = nativeEt;
+        this.isNative = isNative;
         init();
     }
 
@@ -114,6 +116,8 @@ public abstract class NettyTcpConnector extends NettyConnector {
         EventLoopGroup worker = worker();
         if (worker instanceof EpollEventLoopGroup) {
             ((EpollEventLoopGroup) worker).setIoRatio(workerIoRatio);
+        } else if (worker instanceof KQueueEventLoopGroup) {
+            ((KQueueEventLoopGroup) worker).setIoRatio(workerIoRatio);
         } else if (worker instanceof NioEventLoopGroup) {
             ((NioEventLoopGroup) worker).setIoRatio(workerIoRatio);
         }
@@ -121,13 +125,44 @@ public abstract class NettyTcpConnector extends NettyConnector {
 
     @Override
     protected EventLoopGroup initEventLoopGroup(int nThreads, ThreadFactory tFactory) {
-        return isNativeEt() ? new EpollEventLoopGroup(nThreads, tFactory) : new NioEventLoopGroup(nThreads, tFactory);
+        if (isNativeEPoll()) {
+            return new EpollEventLoopGroup(nThreads, tFactory);
+        }
+        if (isNativeKQueue()) {
+            return new KQueueEventLoopGroup(nThreads, tFactory);
+        }
+        return new NioEventLoopGroup(nThreads, tFactory);
     }
 
     /**
-     * Netty provides the native socket transport for Linux using JNI based on Epoll Edge Triggered(ET).
+     * Netty provides the native socket transport for Linux using JNI.
      */
-    public boolean isNativeEt() {
-        return nativeEt && NativeSupport.isSupportNativeET();
+    public boolean isNativeEPoll() {
+        return isNative && NativeSupport.isNativeEPollAvailable();
+    }
+
+    /**
+     * Netty provides the native socket transport for BSD systems such as MacOS using JNI.
+     */
+    public boolean isNativeKQueue() {
+        return isNative && NativeSupport.isNativeKQueueAvailable();
+    }
+
+    protected void initChannelFactory() {
+        if (isNativeEPoll()) {
+            bootstrap().channelFactory(TcpChannelProvider.NATIVE_EPOLL_CONNECTOR);
+        } else if (isNativeKQueue()) {
+            bootstrap().channelFactory(TcpChannelProvider.NATIVE_KQUEUE_CONNECTOR);
+        } else {
+            bootstrap().channelFactory(TcpChannelProvider.NIO_CONNECTOR);
+        }
+    }
+
+    @Override
+    public String toString() {
+        return "isNativeEPoll: " + isNativeEPoll()
+                + ", isNativeKQueue: " + isNativeKQueue()
+                + JConstants.NEWLINE
+                + bootstrap();
     }
 }