+++ release-notes.md
@@ -1,11 +1,13 @@
 Jupiter release notes
 ------------------------
 
-### 2017-05-06:
+### 2017-05-19:
 
 - 存在addAndGet场景时, 不使用jupiter自带的Atomic*FieldUpdater, 使用jdk默认的实现, 1.8后续版本jdk使用了intrinsics后更快
 - 移除org.jupiter.rpc.model.metadata.ServiceWrapper#connCount, connCount修正为JServer粒度的参数
 - 修复权重问题
+- 升级Netty版本到4.1.11.Final
+- netty-transport-native-kqueue 支持 [\#20](https://github.com/fengjiachun/Jupiter/issues/20)
 
 ### 2017-04-18: version 1.2.4